+++ sshd-common/src/main/java/org/apache/sshd/common/util/closeable/AbstractCloseable.java
@@ -26,14 +26,22 @@ import org.apache.sshd.common.future.SshFuture;
 import org.apache.sshd.common.future.SshFutureListener;
 
 /**
- * Provides some default implementations
+ * Provides some default implementations for managing channel/connection open/close state
  *
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
 public abstract class AbstractCloseable extends IoBaseCloseable {
 
     public enum State {
-        Opened, Graceful, Immediate, Closed
+        /** Connection is open */
+        Opened,
+        /** Connection is being closed gracefully */
+        Graceful,
+        /** Connection is being terminated immediately */
+        Immediate,
+        /** Connection is closed */
+        Closed,
+        /* end */;
     }
 
     /**
@@ -90,7 +98,7 @@ public abstract class AbstractCloseable extends IoBaseCloseable {
                 }
             } else {
                 if (debugEnabled) {
-                    log.debug("close({})[Immediately] state already {}", this, state.get());
+                    log.debug("close({})[Immediately] state already {}", this, state);
                 }
             }
         } else {
@@ -119,7 +127,7 @@ public abstract class AbstractCloseable extends IoBaseCloseable {
                 }
             } else {
                 if (debugEnabled) {
-                    log.debug("close({})[Graceful] state already {}", this, state.get());
+                    log.debug("close({})[Graceful] state already {}", this, state);
                 }
             }
         }